third_party/blink/web_tests/fast/workers/worker-sharedarraybuffer-transfer-expected.txt

CONSOLE WARNING: line 113: SharedArrayBuffer will require cross-origin isolation as of M91, around May 2021. See https://developer.chrome.com/blog/enabling-shared-array-buffer/ for more details.
CONSOLE WARNING: line 27: SharedArrayBuffer will require cross-origin isolation as of M91, around May 2021. See https://developer.chrome.com/blog/enabling-shared-array-buffer/ for more details.
Test sharing SharedArrayBuffers between dedicated workers.

Running SharedArrayBuffer test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred data is OK.
Running Int8Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Uint8Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Uint8ClampedArray test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Int16Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Uint16Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Int32Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Uint32Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Float32Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Float64Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running SendSharedArrayBufferTwice test case
PASS: Transferred data is OK.
Running TransferArrayBufferAndSharedArrayBuffer test case
PASS: ArrayBuffer was neutered during transfer.
PASS: Transferred data is OK.
PASS: Transferred data is OK.
Running SharedArrayBufferInTransferListViaMessagePort test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: ArrayBuffer 'ab1' was not neutered.
PASS: ArrayBuffer 'ab2' was not neutered.
DONE
